requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
psycopg2-binary
python-multipart
openai
pillow